+++ libf2c/libI77/uio.c
@@ -10,7 +10,7 @@ do_us (ftnint * number, char *ptr, ftnlen len)
       f__recpos += (int) (*number * len);
       if (f__recpos > f__reclen)
 	err (f__elist->cierr, 110, "do_us");
-      if (fread (ptr, (size_t) len, (size_t) (*number), f__cf) != *number)
+      if (fread (ptr, (size_t) len, (size_t) (*number), f__cf) != (size_t) *number)
 	err (f__elist->ciend, EOF, "do_us");
       return (0);
     }
@@ -34,7 +34,7 @@ do_ud (ftnint * number, char *ptr, ftnlen len)
       if (!(i = fread (ptr, (size_t) len, (size_t) (*number), f__cf))
 	  && !(f__recpos - *number * len))
 	err (f__elist->cierr, EOF, "do_ud");
-      if (i < *number)
+      if (i < (size_t) *number)
 	memset (ptr + i * len, 0, (*number - i) * len);
       return 0;
 #else
